Merge dev-1.50 into main branch

+++ src/main/java/bassebombecraft/operator/entity/RemoveCharm2.java
@@ -1,10 +1,16 @@
 package bassebombecraft.operator.entity;
 
 import static bassebombecraft.BassebombeCraft.getProxy;
+import static bassebombecraft.operator.DefaultPorts.getFnGetLivingEntity1;
+import static bassebombecraft.operator.Operators2.applyV;
 
+import java.util.function.Function;
+
+import bassebombecraft.entity.EntityUtils;
 import bassebombecraft.event.charm.ServerCharmedMobsRepository;
 import bassebombecraft.operator.Operator2;
 import bassebombecraft.operator.Ports;
+import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.MobEntity;
 
 /**
@@ -13,21 +19,48 @@ import net.minecraft.entity.MobEntity;
  * {@linkplain ServerCharmedMobsRepository}.
  * 
  * Sets the result port to true if removal succeeds.
- * 
- * The charm is removed from living entity #1 in the ports.
  */
 public class RemoveCharm2 implements Operator2 {
 
+	/**
+	 * Function to get living entity.
+	 */
+	Function<Ports, LivingEntity> fnGetLivingEntity;
+
+	/**
+	 * Constructor.
+	 * 
+	 * @param fnGetLivingEntity function to get entity to test.
+	 */
+	public RemoveCharm2(Function<Ports, LivingEntity> fnGetLivingEntity) {
+		this.fnGetLivingEntity = fnGetLivingEntity;
+	}
+
+	/**
+	 * Constructor.
+	 * 
+	 * Instance is configured with living entity #1 from ports.
+	 */
+	public RemoveCharm2() {
+		this(getFnGetLivingEntity1());
+	}
+
 	@Override
-	public Ports run(Ports ports) {
-		
+	public void run(Ports ports) {
+		LivingEntity entity = applyV(fnGetLivingEntity, ports);
+
+		// exit if entity isn't a mob entity
+		if (!EntityUtils.isTypeMobEntity(entity))
+			return;
+
+		// type cast
+		MobEntity mobEntity = (MobEntity) entity;
+
 		// remove from repository
-		MobEntity mobEntity = (MobEntity) ports.getLivingEntity1();
 		getProxy().getServerCharmedMobsRepository().remove(mobEntity);
 
 		// update result port
 		ports.setResultAsSucces();
-		return ports;
 	}
 
 }
